Updated Navigation component
